+++ CHANGES.md
@@ -41,6 +41,7 @@ This file is a best-effort approach to solving this issue; we will do our best b
 ### Changed behavior
 
 * (lr-wpan) Association: Fix the handling of situations where the association response commands arrives before the data request command acknowledgment that is supposed to precede it.
+* (wifi) Minstrel and MinstrelHt stats file default name changed to include both local and remote mac addresses (last 5 characters of address used i.e. xx:xx). New format: `minstrel-stats-<local mac address>-<remote mac address>.txt` and `minstrel-ht-stats-<local mac address>-<remote mac address>.txt`.
 
 ## Changes from ns-3.42 to ns-3.43